FLASK/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
psutil
requests
